New summary table creation for minnes task
Minne made more progress on her task, and it turns out it is easier to export csv files with total success and total failures so each participant is in 1 row. This means that my import data and remove errors function are not as needed -- so i made a new script to take minnes output and calculate success rate along with reaction time avg.